EnSys__June8/EnSys-main/requirements.txt
flask
vaderSentiment
spacy
nltk
python-dotenv
-U scikit-learn
pandas
openai==0.28
python -m spacy download en_core_web_sm
python-docx
pip install